[refactor] home 컴포넌트 분리
