+++ src/components/home/AttendanceBox.vue
@@ -1,0 +1,20 @@
+<!-- 팀 근태 현황 박스 -->
+<template>
+  <div class="box attendance">팀 근태 현황</div>
+</template>
+
+<style scoped>
+.attendance {
+  grid-area: attendance;
+  min-height: 420px;
+}
+.box {
+  background: #fff;
+  border-radius: 15px;
+  padding: 20px;
+  min-height: 200px;
+  box-shadow: 1px 1px 20px 1px rgba(0, 0, 0, 0.05);
+  font-size: 17px;
+  font-weight: bold;
+}
+</style>
