+++ src/components/home/MyStatusBox.vue
@@ -1,0 +1,20 @@
+<!-- 나의 현황 박스 -->
+<template>
+  <div class="box my-status">나의 현황</div>
+</template>
+
+<style scoped>
+.my-status {
+  grid-area: my-status;
+  min-height: 400px !important;
+}
+.box {
+  background: #fff;
+  border-radius: 15px;
+  padding: 20px;
+  min-height: 200px;
+  box-shadow: 1px 1px 20px 1px rgba(0, 0, 0, 0.05);
+  font-size: 17px;
+  font-weight: bold;
+}
+</style>
