+++ src/components/home/NoticeBox.vue
@@ -1,0 +1,19 @@
+<!-- 공지사항 박스 -->
+<template>
+  <div class="box notice">공지 사항</div>
+</template>
+
+<style scoped>
+.notice {
+  grid-area: notice;
+}
+.box {
+  background: #fff;
+  border-radius: 15px;
+  padding: 20px;
+  min-height: 200px;
+  box-shadow: 1px 1px 20px 1px rgba(0, 0, 0, 0.05);
+  font-size: 17px;
+  font-weight: bold;
+}
+</style>
